move domain comparision util from constants to utls
